+++ src/model/PhaseConstraint.cpp
@@ -92,7 +92,7 @@ void PhaseConstraint::updateGamma()
         mGamma = mGammaFixed;
     else if(mGammaType == eGammaRange && mGammaMax > mGammaMin)
     {
-        float max = qMin(mGammaMax, mPhaseTo->mAlpha.mX - mPhaseFrom->mBeta.mX);
+        double max = qMin(mGammaMax, mPhaseTo->mAlpha.mX - mPhaseFrom->mBeta.mX);
         mGamma = Generator::randomUniform(mGammaMin, max);
     }
 }
